requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
pytz==2022.7.1
sqlparse==0.4.3
